+++ test/support/schema1.sql
@@ -54,6 +54,11 @@ CREATE TABLE chapters (
   pages INT
 );
 
+DROP TABLE IF EXISTS stores;
+CREATE TABLE stores (
+  name TEXT
+);
+
 DROP TABLE IF EXISTS robots;
 CREATE TABLE robots (
   id SERIAL PRIMARY KEY,