+++ test/sync_test.rb
@@ -56,6 +56,10 @@ class SyncTest < Minitest::Test
     assert_result("--preserve", source, dest, expected, "books")
   end
 
+  def test_all
+    assert_works "", dbs: true
+  end
+
   def test_no_shared_fields
     assert_prints "authors: No fields to copy", "authors", dbs: true
   end